+++ Changelog.md
@@ -1,3 +1,11 @@
+Improved Mobs 1.8.1
+================
+- Make parrot ability get triggered less
+- Tweak some default equipment weights so diamond/netherite is rarer
+- Add quality value to equipment config.  
+  Makes it able to specify which equipment will get more  
+  likely choosen with higher difficulty
+
 Improved Mobs 1.8.0
 ================
 - Port to 1.18